// types.ts
export enum AppMode {
  STOPWATCH = 'STOPWATCH',
  COUNTDOWN = 'COUNTDOWN',
  RACE_TIMERS = 'RACE_TIMERS',
  TIMERS = 'TIMERS',
  NAME_PICKERS = 'NAME_PICKERS',
  RANDOM_NUMBERS = 'RANDOM_NUMBERS',
  SENSORY_TIMERS = 'SENSORY_TIMERS',
  CLOCKS = 'CLOCKS',
  CHANCE_GAMES = 'CHANCE_GAMES',
  GROUP_GENERATOR = 'GROUP_GENERATOR',
  PRESENTATION_TIMERS = 'PRESENTATION_TIMERS'
}

export interface Lap {
  id: number;
  time: number;
  split: number;
}